# docs/swagger.yaml
basePath: /
definitions:
  helper.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: OK
        type: string
    type: object
...
host: localhost:1323
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your-Project Document
  version: "1.0"
paths:
...
  /user/filter/name:
    get:
      consumes:
      - application/json
      description: Get users from the database Filter
      parameters:
      - description: fname
        in: query
        name: fname
        type: string
      - description: lname
        in: query
        name: lname
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
      summary: Get user Filter Name data
      tags:
      - User
...
schemes:
- http
swagger: "2.0"